requirements.txt
fastapi==0.116.1
python-dotenv==1.1.1
SQLAlchemy==2.0.43
strawberry-graphql==0.146.0
uvicorn[standard]==0.23.1
psycopg2-binary
alembic
pyreqs
python-jose
passlib